# application/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "1.0.0",
      "title": "sanjib's code for flask api",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "tags": [
        {
          "name": "Products API",
          "description": "Test API"
        }
      ],
      "servers": [
        {
        "url": "/"
        }
    ],
    "components": {
        "schemas": {
            "Todolist": {
                "type": "object",
                "properties": {
                    "Name": {
                    "type": "string"
                    },
                    "Description": {
                    "type": "string"
                    },
                    "Type": {
                        "type": "string"
                    },
                    "Age": {
                        "type": "integer"
                    }
                }
            },
            "Id":{
            "type" : "integer",
            "properties" : {
                "name" : {
                    "type" : "integer"
                }
            }
        },
        "update":{
            "type": "object",
            "properties" :{
               "Name": {
                  "type": "string"
                  },
                "Description": {
                  "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "Age": {
                    "type": "integer"
                },
                "Checked": {
                  "type": "boolean"
                }
            }
          }
        }
    },
    "paths": {
        "/todolist": {
            "post": {
            "tags": [
                "Add todo to API"
            ],
            "summary": "Add todo",
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
            }
            }
            },
            "get" :{
                "tags": [
                    "Get All todos from API"
                ],
                "summary": "Get All todos",
               
                "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
            }
            }
            }
        },
        "/todolist/{id}":{

            "get" :{
                "tags": [
                    "Get todos from API"
                ],
                "summary": "Get todos",
                "parameters": [{
                  "name": "id",
                  "in": "path",
                  "description": "todo id to update",
                  "required": true,
                  "type": "integer",
                  "schema" : {
                      "$ref": "#/components/schemas/Id"
                  }
                }],
               
                "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
            }
            }
            },
           
            "put": {
            "tags": [
                "Update todos on API"
            ],
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "todo id to update",
                "required": true,
                "type": "integer",
                "schema" : {
                    "$ref": "#/components/schemas/Id"
                }
            }],
            "summary": "Update Todo",
            "requestBody": {
            "description": "Update Existing todo On Database",
            "required": true,
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/update"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
            }
            }
        },
        "delete": {
            "tags": [
                "Delete todo to API"
            ],
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "product id to delete",
                "required": true,
                "type": "integer",
                "schema" : {
                    "$ref": "#/components/schemas/Id"
                }
            }],
            "summary": "Delete todo",
            
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Todolist"
                    }
            }
            }
        }
        }
}

} 